UPD: Fixed type - renamed 'docket_check.py' to 'docker_check.py'
