+++ backend/core/tasks/schedules.py
@@ -8,7 +8,7 @@ from core.tasks.api_refresh import api_refresh
 from core.tasks.cleanup import trailer_cleanup
 from core.tasks.download_trailers import download_missing_trailers
 from core.tasks.image_refresh import refresh_images
-from core.updates.docket_check import check_for_update
+from core.updates.docker_check import check_for_update
 
 # from core.tasks.task_runner import TaskRunner
